Fix bypassing Request Body scanning with trailers
The present implementation of WASM plugin is missing a scenario
that onRequestBody method will never be called with end_of_stream
parameter set to true. Such situation will happen for HTTP 2
request with trailers as the presence of trailers means that the
last chunk of payload is still not the end of the request stream.

This is problematic because existing implementation triggers
threat scanning when onRequestBody is called with parameter
end_of_stream set to true as it was expected to be a proper place:

* right after collecting entire payload

* before completing the request stream

This is true only for scenarios that do NOT include request trailers
(unless it is HTTP1 as Envoy Proxy ignores request trailers for that
protocol - the explanation is that trailers in HTTP1 are very often
not handled properly by various servers even though HTTP standard
allows it).

To fix it, onRequestTrailers method was implemented to cover the
situation when request payload was collected as onRequestTrailers
is the method which applies to the situation:

* right after collecting entire payload

* before completing the request stream

To maintain the existing logic, this method calls onRequestBody
with end_of_stream set to true, to share the same logic.

The onResponseTrailers method was not implemented but perhaps
it should be implemented as well.